requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.1
django-dirtyfields==1.8.2
django-environ==0.9.0
django-filter==22.1
django-flex-user==1.0.1
django-heroku==0.3.1
django-phonenumber-field==6.3.0
django-redis==5.2.0
django-rest-multiple-models==2.1.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
environ==1.0
filelock==3.8.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
oauthlib==3.2.0
packaging==21.3
phonenumbers==8.12.53
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
python-env==1.0.0
python3-openid==3.2.0
pytz==2022.1
redis==4.3.4
regex==2022.7.25
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
split-settings==1.0.0
sqlparse==0.4.2
tldextract==3.3.1
urllib3==1.26.11
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.8.1
